Deployable/usable tiddlywiki shell.
This shell extends the proof of concept by writing the firebase
configuration to the correct spot in localstorage, and serving
all URLs under /tw/[name] as a tiddlywiki. Then the static build
served for /tw/poc for example looks in localstorage for the
firestore config, and if it finds it works against that firestore
instance.

Meanwhile, since the configuration for the wide open proof of
concept is hardcoded into the top level landing page, anyone can
click the proof of concept link and see and later edit the wiki.

Firestore's support for svelte works fine, but I am not sure about
using a cloud function for SSR. It is definitely slow the first time
and doesn't yet provide any advantage.

Firestore also has some sort of node engine mismatch bug which can
be worked around by modifying the .npmrc to not enforce strict
checking that the local node and the node expected by functions
is identical.

But by and large, this is working great and just needs programmatic
creation of firestore instances and the ability to write those to
local storage added to make it a fully functioning wiki creation
and sharing server.